+++ include/envoy/ratelimit/ratelimit.h
@@ -1,16 +1,7 @@
 #pragma once
 
-#include <chrono>
-#include <memory>
 #include <string>
 #include <vector>
-
-#include "envoy/common/pure.h"
-#include "envoy/config/bootstrap/v2/bootstrap.pb.h"
-#include "envoy/singleton/instance.h"
-#include "envoy/tracing/http_tracer.h"
-
-#include "absl/types/optional.h"
 
 namespace Envoy {
 namespace RateLimit {
@@ -29,21 +20,6 @@ struct DescriptorEntry {
 struct Descriptor {
   std::vector<DescriptorEntry> entries_;
 };
-
-/**
- * RateLimitServiceConfig that wraps the proto structure.
- */
-class RateLimitServiceConfig : public Singleton::Instance {
-
-public:
-  RateLimitServiceConfig(
-      const envoy::config::ratelimit::v2::RateLimitServiceConfig& ratelimit_config)
-      : config_(ratelimit_config) {}
-
-  const envoy::config::ratelimit::v2::RateLimitServiceConfig& config_;
-};
-
-typedef std::shared_ptr<RateLimitServiceConfig> RateLimitServiceConfigPtr;
 
 } // namespace RateLimit
 } // namespace Envoy
