+++ include/envoy/server/filter_config.h
@@ -9,7 +9,6 @@
 #include "envoy/json/json_object.h"
 #include "envoy/network/drain_decision.h"
 #include "envoy/network/filter.h"
-#include "envoy/ratelimit/ratelimit.h"
 #include "envoy/runtime/runtime.h"
 #include "envoy/server/admin.h"
 #include "envoy/server/overload_manager.h"
@@ -134,11 +133,6 @@ public:
    * @return OverloadManager& the overload manager for the server.
    */
   virtual OverloadManager& overloadManager() PURE;
-
-  /**
-   * @return RateLimit::RateLimitSErviceConfigPtr for use by extensions.
-   */
-  virtual RateLimit::RateLimitServiceConfigPtr rateLimitServiceConfig() const PURE;
 };
 
 class ListenerFactoryContext : public FactoryContext {
