+++ source/extensions/filters/common/ratelimit/ratelimit_impl.cc
@@ -14,6 +14,8 @@
 #include "common/http/header_map_impl.h"
 #include "common/http/headers.h"
 
+#include "server/configuration_impl.h"
+
 namespace Envoy {
 namespace Extensions {
 namespace Filters {
@@ -26,8 +28,8 @@ SINGLETON_MANAGER_REGISTRATION(ratelimit_client);
 ClientFactoryPtr
 ClientFactory ::rateLimitClientFactory(Server::Configuration::FactoryContext& context) {
   // TODO(ramaraochavali): figure out a way to get the singleton name here.
-  Envoy::RateLimit::RateLimitServiceConfigPtr ratelimit_config =
-      context.singletonManager().getTyped<Envoy::RateLimit::RateLimitServiceConfig>(
+  Envoy::Server::Configuration::RateLimitServiceConfigPtr ratelimit_config =
+      context.singletonManager().getTyped<Envoy::Server::Configuration::RateLimitServiceConfig>(
           "ratelimit_config_singleton", [] { return nullptr; });
   if (ratelimit_config) {
     return std::make_unique<Envoy::Extensions::Filters::Common::RateLimit::GrpcFactoryImpl>(
