+++ source/server/configuration_impl.cc
@@ -57,12 +57,10 @@ void MainImpl::initialize(const envoy::config::bootstrap::v2::Bootstrap& bootstr
   }
 
   if (bootstrap.has_rate_limit_service()) {
-    ratelimit_service_config_ =
-        server.singletonManager().getTyped<Envoy::RateLimit::RateLimitServiceConfig>(
-            SINGLETON_MANAGER_REGISTERED_NAME(ratelimit_config), [&bootstrap] {
-              return std::make_shared<Envoy::RateLimit::RateLimitServiceConfig>(
-                  bootstrap.rate_limit_service());
-            });
+    ratelimit_service_config_ = server.singletonManager().getTyped<RateLimitServiceConfig>(
+        SINGLETON_MANAGER_REGISTERED_NAME(ratelimit_config), [&bootstrap] {
+          return std::make_shared<RateLimitServiceConfig>(bootstrap.rate_limit_service());
+        });
   }
 
   ENVOY_LOG(info, "loading {} cluster(s)", bootstrap.static_resources().clusters().size());
