+++ test/extensions/filters/http/ratelimit/config_test.cc
@@ -16,8 +16,6 @@ namespace Extensions {
 namespace HttpFilters {
 namespace RateLimitFilter {
 
-SINGLETON_MANAGER_REGISTRATION(ratelimit_config); // just of tests.
-
 TEST(RateLimitFilterConfigTest, ValidateFail) {
   NiceMock<Server::Configuration::MockFactoryContext> context;
   EXPECT_THROW(RateLimitFilterConfig().createFilterFactoryFromProto(
